requirements.txt
pillow
ipython
ipykernel
pandas
pdfkit
fastapi
uvicorn